+++ InventorySystem_Project/src/main/java/com/inventorysystem_project/serviceinterfaces/IRolService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface IRolService {
     public void insert(Rol rol);
     public List<Rol> list();
-    public void delete(int idRol);
-    public Rol listId(int idRol);
+    public void delete(Long idRol);
+    public Rol listId(Long idRol);
 }
